Redirecting the User to the Checkout Page
When the user starts to build a burger from the root page, after the user logged in successfully set the path to '/checkout' for redirecting.

Add a state to burger builder reducer, for storing the user did choose the ingredients or not.

Add a state to Auth reducer for setting redirect path after authenticated.

After logging in, redirect by path of state from Auth page to root page if haven't build a burger yet, or continue to the Checkout page if already builds a burger.