Add search title using useEffect and useCallback
To avoid  unnecessary re-renders and an infinite loop, ensure we don't have to do extra render a cycle initially.

Get rid of useEffect hook in Ingredients component, because we are fetching ingredients with search component.

Using useCallback to avoid the functions defined inside of the component function are created again.