Add update ingredient and Http state using useReducer
Use Reducer to manage to state that is related to that request being sent that influences are a UI.

Take some logic out of Ingredients component and move it into the reducer functions.

So we have cleaner code where we have a clearly defined flow of data and where we change our data.