Optimizing the Custom Store Hook
Problem:
When click a button, it gets re-rendered four times because I have four products.
But actually should only render one time because this clicked item changed, we're updating the button look and so on.
The other three items shouldn't really be rendered.

Optimize:
Add React.memo to Product Item component.
Add an argument to determine whether we actually want to register a listener to this component or not.